+++ sw.js
@@ -1,8 +1,8 @@
-const CACHE='kirin-anikku-v015';
+const CACHE='kirin-anikku-v020';
 const SHELL=[
-  './','./index.html','./assets/css/app.css?v=015','./assets/js/app.js?v=015',
+  './','./index.html','./assets/css/app.css?v=020','./assets/js/app.js?v=020',
   './assets/vendor/pako.min.js','./assets/icons/app-icon.svg','./schemas/schema-anikku.proto',
-  './manifest.webmanifest?v=015'
+  './manifest.webmanifest?v=020'
 ];
 const CDN=[
   'https://cdn.jsdelivr.net/npm/long@5.2.3/umd/index.min.js',